+++ frontend/src/assets/liclogo.png
