+++ frontend/src/assets/policy.jpg
