+++ frontend/src/assets/starlogo.png
