+++ frontend/src/components/CarouselScroll.tsx
