+++ frontend/src/components/NavBar.tsx
